requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20

# Config
python-dotenv==1.0.1

# MQTT
paho-mqtt==2.1.0

# Data Processing
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1

# Database
influxdb-client==1.48.0
psycopg2-binary==2.9.10

# Notification
python-telegram-bot==21.10

# Testing
pytest==8.3.4
httpx==0.28.1
